+++ data-raw/package-data-related/code/generate-package-internal-data.R
@@ -1,0 +1,10 @@
+# Source this file from PCAWG7 package root
+
+cat(getwd(), "\n")
+
+source("data-raw/package-data-related/code/convert-tcga_id-to-icgc_specimen_id.R")
+source("data-raw/package-data-related/code/create-sigs-etiology-data.R")
+
+usethis::use_data(TCGA.ICGC.IDs, sigs.etiologies,
+                  internal = TRUE, overwrite = TRUE)
+