+++ tests/testthat.R
@@ -1,0 +1,5 @@
+library(testthat)
+library(PCAWG7)
+
+test_check("PCAWG7")
+