+++ tests/testthat/test_SplitPCAWGMatrixByTumorType.R
@@ -1,0 +1,4 @@
+test_that("SplitPCAWGMatrixByTumorType", {
+  spectra.list <- SplitPCAWGMatrixByTumorType(spectra$PCAWG$SBS96)
+  expect_equal(length(spectra.list), 37)
+})